refactor(substrate): extract a shared atomic_write helper
The three near-identical atomic file-write functions — in pid_lock,
handle_store, and aether-capabilities/store — collapse into a single
`aether_substrate::atomic_write::atomic_write`. Each site now imports
and calls the shared helper; the nonce switches from per-site millis/nanos
sources to a consistent nanosecond value in the new module. The per-file
`now_millis` / `now_nanos` helpers (used for other purposes in their
respective files) are kept.

The redundant path prefix in the `acquire_lock_pid` intra-doc link in
handle_store.rs is also dropped.

These resolve the Qodana duplicate-code and redundant-qualifier findings
on PR #1970.
